requirements.txt
# Core Flask dependencies
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.6.0
Flask-CORS==5.0.0

# Database
SQLAlchemy==2.0.41
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Security
Werkzeug==3.0.6
bcrypt==4.3.0
PyJWT==2.9.0

# Communication services
sendgrid==6.10.0  # Email notifications
twilio==9.6.3     # SMS notifications

# Environment and utilities
python-dotenv==1.0.1
python-dateutil==2.9.0.post0

# Testing dependencies
pytest==8.3.5
pytest-flask==1.3.0

# Development dependencies
alembic==1.14.1

# Core Python utilities
requests==2.32.3
urllib3==2.2.3
certifi==2025.4.26

# Template and markup
Jinja2==3.1.6
MarkupSafe==2.1.5

# Other utilities
click==8.1.8
itsdangerous==2.2.0
blinker==1.8.2
importlib-metadata==8.5.0
zipp==3.20.2

# Optional: For production deployment
gunicorn==21.2.0  # WSGI server for production
